requirements.txt
requests==2.31.0
pandas==2.1.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
python-dotenv==1.0.0
pytz==2023.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
colorama==0.4.6
tabulate==0.9.0